+++ AuthPermissions.AspNetCore/Services/IShardingConnections.cs
@@ -18,10 +18,10 @@ public interface IShardingConnections
     IEnumerable<string> GetAllConnectionStringNames();
 
     /// <summary>
-    /// This returns all the connection string names, with the number of tenants linked to those connection string names
+    /// This returns all the connection string names, with a list of tenant name  linked to each connection name
     /// </summary>
-    /// <returns>List of KeyValuePair(string, int) in the order they are in the appsettings file</returns>
-    Task<List<KeyValuePair<string, int>>> GetConnectionStringsWithNumTenantsAsync();
+    /// <returns>List of connectionName+tenants using that connectionName</returns>
+    Task<List<(string connectionName, List<string> tenantNames)>> GetConnectionStringsWithTenantNamesAsync();
 
     /// <summary>
     /// This will provide the connection string for the entry with the given connection string name